requirements.txt
discord.py
discord.py[voice]
PyNaCl
nltk
tswift
google-cloud-texttospeech
pydub
google-auth
